backend/requirements.txt
fastapi
uvicorn
scikit-learn
joblib
pandas
numpy
python-multipart
pdfplumber
python-docx

# uvicorn app:app --reload